chore: move update review to reviews router
